requirements.txt
streamlit
openai
python-dotenv
pandas
fpdf
pydrive
openpyxl